requirements.txt
google-genai
Flask
requests
gunicorn  # 新增 gunicorn